Separate UI template files for further enhancement.
