+++ oap-server/server-bootstrap/pom.xml
@@ -257,14 +257,11 @@
                         <exclude>gateways.yml</exclude>
                         <exclude>service-apdex-threshold.yml</exclude>
                         <exclude>endpoint-name-grouping.yml</exclude>
-                        <exclude>ui-initialized-templates.yml</exclude>
-                        <exclude>oal/browser.oal</exclude>
-                        <exclude>oal/core.oal</exclude>
-                        <exclude>oal/java-agent.oal</exclude>
-                        <exclude>oal/dotnet-agent.oal</exclude>
-                        <exclude>oal/envoy.oal</exclude>
+                        <exclude>oal/</exclude>
                         <exclude>fetcher-prom-rules/</exclude>
                         <exclude>meter-receive-config/</exclude>
+                        <exclude>oc-rules/</exclude>
+                        <exclude>ui-initialized-templates/</exclude>
                     </excludes>
                 </configuration>
             </plugin>
