+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModuleProvider.java
@@ -18,6 +18,8 @@
 
 package org.apache.skywalking.oap.server.core;
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import org.apache.skywalking.oap.server.configuration.api.ConfigurationModule;
@@ -145,14 +147,14 @@ public class CoreModuleProvider extends ModuleProvider {
         }
         EndpointNameGrouping endpointNameGrouping = new EndpointNameGrouping();
         this.registerServiceImplementation(NamingControl.class, new NamingControl(
-                moduleConfig.getServiceNameMaxLength(),
-                moduleConfig.getInstanceNameMaxLength(),
-                moduleConfig.getEndpointNameMaxLength(),
-                endpointNameGrouping
+            moduleConfig.getServiceNameMaxLength(),
+            moduleConfig.getInstanceNameMaxLength(),
+            moduleConfig.getEndpointNameMaxLength(),
+            endpointNameGrouping
         ));
         try {
             endpointNameGroupingRuleWatcher = new EndpointNameGroupingRuleWatcher(
-                    this, endpointNameGrouping);
+                this, endpointNameGrouping);
         } catch (FileNotFoundException e) {
             throw new ModuleStartException(e.getMessage(), e);
         }
@@ -180,8 +182,8 @@ public class CoreModuleProvider extends ModuleProvider {
 
         if (moduleConfig.isGRPCSslEnabled()) {
             grpcServer = new GRPCServer(moduleConfig.getGRPCHost(), moduleConfig.getGRPCPort(),
-                    moduleConfig.getGRPCSslCertChainPath(),
-                    moduleConfig.getGRPCSslKeyPath()
+                                        moduleConfig.getGRPCSslCertChainPath(),
+                                        moduleConfig.getGRPCSslKeyPath()
             );
         } else {
             grpcServer = new GRPCServer(moduleConfig.getGRPCHost(), moduleConfig.getGRPCPort());
@@ -217,7 +219,7 @@ public class CoreModuleProvider extends ModuleProvider {
 
         this.registerServiceImplementation(ConfigService.class, new ConfigService(moduleConfig));
         this.registerServiceImplementation(
-                DownSamplingConfigService.class, new DownSamplingConfigService(moduleConfig.getDownsampling()));
+            DownSamplingConfigService.class, new DownSamplingConfigService(moduleConfig.getDownsampling()));
 
         this.registerServiceImplementation(GRPCHandlerRegister.class, new GRPCHandlerRegisterImpl(grpcServer));
         this.registerServiceImplementation(JettyHandlerRegister.class, new JettyHandlerRegisterImpl(jettyServer));
@@ -236,7 +238,7 @@ public class CoreModuleProvider extends ModuleProvider {
         this.registerServiceImplementation(ModelManipulator.class, storageModels);
 
         this.registerServiceImplementation(
-                NetworkAddressAliasCache.class, new NetworkAddressAliasCache(moduleConfig));
+            NetworkAddressAliasCache.class, new NetworkAddressAliasCache(moduleConfig));
 
         this.registerServiceImplementation(TopologyQueryService.class, new TopologyQueryService(getManager()));
         this.registerServiceImplementation(MetricsMetadataQueryService.class, new MetricsMetadataQueryService());
@@ -251,9 +253,9 @@ public class CoreModuleProvider extends ModuleProvider {
 
         // add profile service implementations
         this.registerServiceImplementation(
-                ProfileTaskMutationService.class, new ProfileTaskMutationService(getManager()));
+            ProfileTaskMutationService.class, new ProfileTaskMutationService(getManager()));
         this.registerServiceImplementation(
-                ProfileTaskQueryService.class, new ProfileTaskQueryService(getManager(), moduleConfig));
+            ProfileTaskQueryService.class, new ProfileTaskQueryService(getManager(), moduleConfig));
         this.registerServiceImplementation(ProfileTaskCache.class, new ProfileTaskCache(getManager(), moduleConfig));
 
         this.registerServiceImplementation(CommandService.class, new CommandService(getManager()));
@@ -265,14 +267,16 @@ public class CoreModuleProvider extends ModuleProvider {
 
         if (moduleConfig.isGRPCSslEnabled()) {
             this.remoteClientManager = new RemoteClientManager(getManager(), moduleConfig.getRemoteTimeout(),
-                    moduleConfig.getGRPCSslTrustedCAPath());
+                                                               moduleConfig.getGRPCSslTrustedCAPath()
+            );
         } else {
             this.remoteClientManager = new RemoteClientManager(getManager(), moduleConfig.getRemoteTimeout());
         }
         this.registerServiceImplementation(RemoteClientManager.class, remoteClientManager);
 
         // Management
-        this.registerServiceImplementation(UITemplateManagementService.class, new UITemplateManagementService(getManager()));
+        this.registerServiceImplementation(
+            UITemplateManagementService.class, new UITemplateManagementService(getManager()));
 
         MetricsStreamProcessor.getInstance().setEnableDatabaseSession(moduleConfig.isEnableDatabaseSession());
         TopNStreamProcessor.getInstance().setTopNWorkerReportCycle(moduleConfig.getTopNReportPeriod());
@@ -296,23 +300,23 @@ public class CoreModuleProvider extends ModuleProvider {
         Address gRPCServerInstanceAddress = new Address(moduleConfig.getGRPCHost(), moduleConfig.getGRPCPort(), true);
         TelemetryRelatedContext.INSTANCE.setId(gRPCServerInstanceAddress.toString());
         if (CoreModuleConfig.Role.Mixed.name()
-                .equalsIgnoreCase(
-                        moduleConfig.getRole())
-                || CoreModuleConfig.Role.Aggregator.name()
-                .equalsIgnoreCase(
-                        moduleConfig.getRole())) {
+                                       .equalsIgnoreCase(
+                                           moduleConfig.getRole())
+            || CoreModuleConfig.Role.Aggregator.name()
+                                               .equalsIgnoreCase(
+                                                   moduleConfig.getRole())) {
             RemoteInstance gRPCServerInstance = new RemoteInstance(gRPCServerInstanceAddress);
             this.getManager()
-                    .find(ClusterModule.NAME)
-                    .provider()
-                    .getService(ClusterRegister.class)
-                    .registerRemote(gRPCServerInstance);
+                .find(ClusterModule.NAME)
+                .provider()
+                .getService(ClusterRegister.class)
+                .registerRemote(gRPCServerInstance);
         }
 
         DynamicConfigurationService dynamicConfigurationService = getManager().find(ConfigurationModule.NAME)
-                .provider()
-                .getService(
-                        DynamicConfigurationService.class);
+                                                                              .provider()
+                                                                              .getService(
+                                                                                  DynamicConfigurationService.class);
         dynamicConfigurationService.registerConfigChangeWatcher(apdexThresholdConfig);
         dynamicConfigurationService.registerConfigChangeWatcher(endpointNameGroupingRuleWatcher);
     }
@@ -335,11 +339,15 @@ public class CoreModuleProvider extends ModuleProvider {
         CacheUpdateTimer.INSTANCE.start(getManager(), moduleConfig.getMetricsDataTTL());
 
         try {
-            new UITemplateInitializer(ResourceUtils.read("ui-initialized-templates.yml"))
+            final File[] templateFiles = ResourceUtils.getPathFiles("ui-initialized-templates");
+            for (final File templateFile : templateFiles) {
+                new UITemplateInitializer(new FileInputStream(templateFile))
                     .read()
                     .forEach(uiTemplate -> {
                         ManagementStreamProcessor.getInstance().in(uiTemplate);
                     });
+            }
+
         } catch (FileNotFoundException e) {
             throw new ModuleStartException(e.getMessage(), e);
         }
@@ -347,9 +355,9 @@ public class CoreModuleProvider extends ModuleProvider {
 
     @Override
     public String[] requiredModules() {
-        return new String[]{
-                TelemetryModule.NAME,
-                ConfigurationModule.NAME
+        return new String[] {
+            TelemetryModule.NAME,
+            ConfigurationModule.NAME
         };
     }
 }
