+++ oap-server/server-core/src/test/resources/test-ui-templates/ui-initialized-templates.yml
