--- oap-server/server-core/src/test/resources/ui-initialized-templates-conflict.yml
@@ -1,51 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# UI templates initialized file includes the default template when the SkyWalking OAP starts up at the first time.
-#
-# Also, SkyWalking would detect the existing templates in the database, once they are missing, all templates in this file
-# could be added automatically.
-
-templates:
-  - name: "APM"
-    # The type includes DASHBOARD, TOPOLOGY_INSTANCE, TOPOLOGY_ENDPOINT.
-    # DASHBOARD type templates could have multiple definitions, by using different names.
-    # TOPOLOGY_INSTANCE, TOPOLOGY_ENDPOINT type templates should be defined once, as they are used in the topology page only.
-    type: "DASHBOARD"
-    # Configuration could be defined through UI, and use `export` to format in the standard JSON.
-    configuration: |-
-      ui exported definition in JSONsdfsd
-      sdjfkasld
-      sdfui
-      skdfj:dafjisa
-      adifaosi
-    # Activated as the DASHBOARD type, makes this templates added into the UI page automatically.
-    # False means providing a basic template, user needs to add it manually.
-    activated: true
-    # True means wouldn't show up on the dashboard. Only keeps the definition in the storage.
-    # disabled: false
-  - name: "APM"
-    # The type includes DASHBOARD, TOPOLOGY_INSTANCE, TOPOLOGY_ENDPOINT.
-    # DASHBOARD type templates could have multiple definitions, by using different names.
-    # TOPOLOGY_INSTANCE, TOPOLOGY_ENDPOINT type templates should be defined once, as they are used in the topology page only.
-    type: "DASHBOARD"
-    # Configuration could be defined through UI, and use `export` to format in the standard JSON.
-    configuration: |-
-      Mesh metrics configuration.
-    # Activated as the DASHBOARD type, makes this templates added into the UI page automatically.
-    # False means providing a basic template, user needs to add it manually.
-    # activated: true
-    # True means wouldn't show up on the dashboard. Only keeps the definition in the storage.
-    disabled: true
